requirements.txt
# Core PyTorch dependencies
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Audio processing
librosa>=0.10.1
soundfile>=0.12.1

# Data science and ML
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Data visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars and utilities
tqdm>=4.65.0

# Configuration and logging
pyyaml>=6.0
tensorboard>=2.13.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional: For enhanced audio processing
scipy>=1.11.0
